feat(console): update connector icons
